Version cleaned up and ready for testing.

+++ axi_simple_dma_prj1/hdl/hls/src/dma_xor_engine.cpp
@@ -4,7 +4,7 @@
 
 #define BLOCK_SIZE 256
 
-void dma1(volatile int* a, unsigned long count, uint32_t key) {
+void dma1(volatile uint32_t * a, uint32_t count, uint32_t key) {
 
 #pragma HLS INTERFACE m_axi depth = 1024 offset = direct bundle = gmem0 port = a 
 #pragma HLS INTERFACE s_axilite port = count bundle = control
@@ -12,7 +12,7 @@ void dma1(volatile int* a, unsigned long count, uint32_t key) {
 
 
     int i;
-    int buff[BLOCK_SIZE];
+    uint32_t buff[BLOCK_SIZE];
 
     // memcpy creates a burst access to memory
     // multiple calls of memcpy cannot be pipelined and will be scheduled
@@ -22,11 +22,11 @@ void dma1(volatile int* a, unsigned long count, uint32_t key) {
     unsigned long offset = 0;
     while(todo) {
         int curlen = BLOCK_SIZE < todo ? BLOCK_SIZE : todo;
-        memcpy(buff, (const int*)(a+offset), curlen * sizeof(int));
+        memcpy(buff, (const uint32_t *)(a+offset), curlen * sizeof(uint32_t));
         for (i = 0; i < curlen; i++) {
             buff[i] = buff[i] ^ key;
         }
-        memcpy((int*)(a+offset), buff, curlen * sizeof(int));
+        memcpy((uint32_t *)(a+offset), buff, curlen * sizeof(uint32_t));
         todo -= curlen;
         offset += curlen;
     }
